Refactor quenching kernels
